requirements.txt
brotli==1.1.0
ConfigParser==7.1.0
cryptography==44.0.0
Cython==3.0.11
dl==0.1.0
Flask==3.1.0
Flask_Cors==5.0.0
HTMLParser==0.0.2
ipaddr==2.2.0
keyring==25.6.0
lxml==5.3.0
numpy==2.2.1
ordereddict==1.1
Pillow==11.0.0
protobuf==5.29.2
pyOpenSSL==24.3.0
tensorflow==2.16.2
thread==2.0.5
wincertstore==0.2.1
xmlrpclib==1.0.1
